[refactor] Exception 패키지 추가 및 이동
