+++ src/main/java/com/appcatalog/error/exception/DataConflictException.java
@@ -1,4 +1,4 @@
-package com.appcatalog.error;
+package com.appcatalog.error.exception;
 
 public class DataConflictException extends RuntimeException {
   public DataConflictException(String message) {
